don't lazy load blocks
