+++ js/app/packages/block-call/definition.ts
@@ -1,17 +1,14 @@
 import { defineBlock, type ExtractLoadType, LoadErrors } from '@core/block';
 import { ENABLE_CALLS } from '@core/constant/featureFlags';
 import { ok } from 'neverthrow';
-import { lazy } from 'solid-js';
+
+import { CallBlockAdapter } from './component/CallBlockAdapter';
 
 export const definition = defineBlock({
   name: 'call',
   description: '',
   defaultFilename: 'Call',
-  component: lazy(() =>
-    import('./component/CallBlockAdapter').then((module) => ({
-      default: module.CallBlockAdapter,
-    }))
-  ),
+  component: CallBlockAdapter,
   async load(source, _intent) {
     if (!ENABLE_CALLS()) return LoadErrors.MISSING;
     if (source.type === 'dss') {
