+++ js/app/packages/block-chat/definition.ts
@@ -4,7 +4,7 @@ import { cognitionApiServiceClient } from '@service-cognition/client';
 import type { Entity } from '@service-cognition/generated/schemas/entity';
 import type { DocumentMetadata } from '@service-storage/generated/schemas/documentMetadata';
 import { ok } from 'neverthrow';
-import { lazy } from 'solid-js';
+import BlockChat from './component/Block';
 
 export const DEFAULT_CHAT_NAME = 'New Chat';
 
@@ -14,7 +14,7 @@ export const definition = defineBlock({
   name: 'chat',
   description: '',
   defaultFilename: DEFAULT_CHAT_NAME,
-  component: lazy(() => import('./component/Block')),
+  component: BlockChat,
   liveTrackingEnabled: true,
   async load(source, intent) {
     if (source.type === 'dss') {
