+++ js/app/packages/block-company/definition.ts
@@ -1,15 +1,12 @@
 import { defineBlock, type ExtractLoadType, LoadErrors } from '@core/block';
 import { ok } from 'neverthrow';
-import { lazy } from 'solid-js';
+
+import { CompanyBlockAdapter } from './component/CompanyBlockAdapter';
 
 export const definition = defineBlock({
   name: 'company',
   description: 'View a CRM company',
-  component: lazy(() =>
-    import('./component/CompanyBlockAdapter').then((module) => ({
-      default: module.CompanyBlockAdapter,
-    }))
-  ),
+  component: CompanyBlockAdapter,
   liveTrackingEnabled: false,
   async load(source, _intent) {
     if (source.type === 'dss') {
