+++ js/app/packages/block-contact/definition.ts
@@ -1,15 +1,12 @@
 import { defineBlock, type ExtractLoadType, LoadErrors } from '@core/block';
 import { ok } from 'neverthrow';
-import { lazy } from 'solid-js';
+
+import { ContactBlockAdapter } from './component/ContactBlockAdapter';
 
 export const definition = defineBlock({
   name: 'contact',
   description: 'View a CRM contact',
-  component: lazy(() =>
-    import('./component/ContactBlockAdapter').then((module) => ({
-      default: module.ContactBlockAdapter,
-    }))
-  ),
+  component: ContactBlockAdapter,
   liveTrackingEnabled: false,
   async load(source, _intent) {
     if (source.type === 'dss') {
