+++ js/app/packages/block-md/definition.ts
@@ -13,8 +13,9 @@ import { storageServiceClient } from '@service-storage/client';
 import type { AccessLevel } from '@service-storage/generated/schemas/accessLevel';
 import type { DocumentMetadata } from '@service-storage/generated/schemas/documentMetadata';
 import { makeFileFromBlob } from '@service-storage/util/makeFileFromBlob';
+import { createSyncServiceSource } from '@service-sync/source';
 import { err, ok } from 'neverthrow';
-import { lazy } from 'solid-js';
+import MarkdownBlock from './component/Block';
 import type { MarkdownRewriteOutput } from './signal/rewriteSignal';
 
 export const definition = defineBlock({
@@ -25,7 +26,7 @@ export const definition = defineBlock({
     { name: 'task', defaultFileName: 'New Task' },
     { name: 'snippet', defaultFileName: 'New Snippet' },
   ],
-  component: lazy(() => import('./component/Block')),
+  component: MarkdownBlock,
   accepted: {
     md: 'text/markdown',
   },
@@ -99,7 +100,6 @@ export const definition = defineBlock({
         }
       }
 
-      const { createSyncServiceSource } = await import('@service-sync/source');
       const { source: syncSource, doInitialSync } = createSyncServiceSource(
         source.id,
         token
