+++ js/app/packages/block-pdf/signal/location.ts
@@ -1,4 +1,3 @@
-import { URL_PARAMS } from '@block-pdf/constants';
 import type { PDFViewer } from '@block-pdf/PdfViewer';
 import {
   FindState,
@@ -31,7 +30,18 @@ import {
   viewerReadySignal,
 } from './pdfViewer';
 
-export { URL_PARAMS };
+export const URL_PARAMS = {
+  pageNumber: 'pdf_page_number',
+  yPos: 'pdf_page_y',
+  x: 'pdf_page_x',
+  width: 'pdf_width',
+  height: 'pdf_height',
+  annotationId: 'pdf_ann_id',
+  searchPage: 'pdf_search_page',
+  searchSnippet: 'pdf_search_snippet',
+  searchRawQuery: 'pdf_search_raw_query',
+  searchHighlightTerms: 'pdf_search_highlight_terms',
+} as const;
 
 export type LocationSearchParams = {
   annotationId?: string;
