+++ js/app/packages/block-pdf/type/placeables.ts
@@ -202,9 +202,3 @@ export const IPlaceableSchema = PlaceableBaseSchema.and(IPayloadSchema).and(
   z.object({ internalId: z.string() })
 );
 export const IPlaceableServerSchema = PlaceableBaseSchema.and(IPayloadSchema);
-
-export function isThreadPlaceable(
-  placeable: IPlaceable
-): placeable is IThreadPlaceable {
-  return placeable.payloadType === 'thread';
-}
