+++ js/app/packages/core/component/LexicalMarkdown/citationsUtils.ts
@@ -1,5 +1,5 @@
 import { URL_PARAMS as MD_URL_PARAMS } from '@block-md/constants';
-import { URL_PARAMS as PDF_URL_PARAMS } from '@block-pdf/constants';
+import { URL_PARAMS as PDF_URL_PARAMS } from '@block-pdf/signal/location';
 import { itemToBlockName } from '@core/constant/allBlocks';
 import { useChannelsContext } from '@core/context/channels';
